semi_data/dataproj/requirements.txt
Django
djangorestframework
mysqlclient
django-concurrency
requests
djangogrpcframework
grpcio
grpcio-tools
googleapis-common-protos